Add errors to vec with new optional function structure
This is a compromise because I was having mutable reference issues before, and needed to take it out of the functions.

I don't like the fact that you always have to check the Option when you use it now, and would prefer it being abstracted to a function. Would be easier to read, but then there'd just be a wrapper function around each goddamn function. Will mess with now, but give up on for now since this is functional.